+++ services/rds/src/main/java/com/g42cloud/sdk/rds/v3/model/RestorePoint.java
@@ -31,7 +31,7 @@ public class RestorePoint {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "restore_time")
 
-    private Integer restoreTime;
+    private Long restoreTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "database_name")
@@ -77,16 +77,16 @@ public class RestorePoint {
         this.backupId = backupId;
     }
 
-    public RestorePoint withRestoreTime(Integer restoreTime) {
+    public RestorePoint withRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
         return this;
     }
 
-    public Integer getRestoreTime() {
+    public Long getRestoreTime() {
         return restoreTime;
     }
 
-    public void setRestoreTime(Integer restoreTime) {
+    public void setRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
     }
