+++ services/rds/src/main/java/com/g42cloud/sdk/rds/v3/model/RestoreToExistingInstanceRequestBodySource.java
@@ -108,7 +108,7 @@ public class RestoreToExistingInstanceRequestBodySource {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "restore_time")
 
-    private Integer restoreTime;
+    private Long restoreTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "database_name")
@@ -154,16 +154,16 @@ public class RestoreToExistingInstanceRequestBodySource {
         this.backupId = backupId;
     }
 
-    public RestoreToExistingInstanceRequestBodySource withRestoreTime(Integer restoreTime) {
+    public RestoreToExistingInstanceRequestBodySource withRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
         return this;
     }
 
-    public Integer getRestoreTime() {
+    public Long getRestoreTime() {
         return restoreTime;
     }
 
-    public void setRestoreTime(Integer restoreTime) {
+    public void setRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
     }
